pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "SemT_py"
version = "0.1"
description = "A utility package for Semantic Enrichment of Tables"
readme = "README.md"
authors = [
    { name = "Alidu Abubakari", email = "a.abubakari@campus.unimib.it" }
]
license = { text = "MIT" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas",
    "numpy",
    "chardet",
    "PyJWT",
    "fake-useragent",
    "requests",
    "python-dateutil",
]

[project.urls]
Homepage = "https://github.com/unimib-datAI/Semtui-python.git"